// src/app/modules/models/project.model.ts
export class Project{
  id:number;
  company:number;
  type: string;
  category:  string;
  internalCode: string;
  description: string;
  idCustomer:  string;
  status: string;
  idSupervisor:number;
  datePlanStart: string;
  datePlanEnd:  string;
  dateStart:  string;
  dateEnd:  string;
  dateDelivery: string;
  workingMinutes:number;
  intervalMinutes:number;
  attachData:[]
}